aggiunto l'endpoint per recuperare gli operatori in base al servizio

+++ src/main/java/org/aldo/beautycenter/data/dto/ServiceDto.java
@@ -8,5 +8,5 @@ public  class ServiceDto {
     private String imgUrl;
     private String name;
     private Double price;
-    private Double duration;
+    private Long duration;
 }
